+++ src/mjlab/tasks/latenttracking/config/__init__.py
@@ -1,0 +1,1 @@
+"""Latent tracking task configurations."""