+++ src/mjlab/tasks/latenttracking/config/g1/__init__.py
@@ -1,0 +1,13 @@
+from mjlab.tasks.latenttracking.rl import LatentTrackingOnPolicyRunner
+from mjlab.tasks.registry import register_mjlab_task
+
+from .env_cfgs import unitree_g1_flat_latent_tracking_bfm_1stage_env_cfg
+from .rl_cfg import unitree_g1_latent_trackingbfm_ppo_runner_cfg
+
+register_mjlab_task(
+  task_id="Mjlab-LatentTrackingbfm-Flat-Unitree-G1-1Stage",
+  env_cfg=unitree_g1_flat_latent_tracking_bfm_1stage_env_cfg(),
+  play_env_cfg=unitree_g1_flat_latent_tracking_bfm_1stage_env_cfg(play=True),
+  rl_cfg=unitree_g1_latent_trackingbfm_ppo_runner_cfg(),
+  runner_cls=LatentTrackingOnPolicyRunner,
+)